Move the interrupt mask primitives from the AES to the BIOS
The cartridge configuration did not link:

    obj/vdi_misc.o: undefined reference to `_enable_interrupts'
    obj/vdi_mouse.o: undefined reference to `_disable_interrupts'

disable_interrupts() and enable_interrupts() were defined in
aes/arch/m68k/gemdosif.S and aes/arch/arm/gemdosifc.c, and aes/ is only
built when CONF_WITH_AES is set -- which the diagnostic cartridge, with no
room for a desktop, leaves unset.  vdi/ is built either way, and it calls
them from vdi_misc.c and vdi_mouse.c.

Upstream EmuTOS did not have this problem: those call sites used set_sr(),
which comes from the BIOS.  set_sr() writes the m68k status register and has
no meaning on ARM, so the port replaced the calls -- the originals are still
there, commented out just above -- and reached for the AES pair without
noticing which layer it was reaching into.  The primitives are not AES
specific in the first place, so move them to the BIOS, which is always
built, following the existing per-architecture split of processor.S and
processor.c.

The declarations in include/arch/m68k/asm.h and include/arch/arm/asm.h stay
where they are; only the definitions move.  gemdosif.S keeps calling them
and now imports them instead of defining them.  The single save slot, and
therefore the fact that these do not nest, is unchanged.

Verified by building the Raspberry Pi configurations with the AES both on
and off -- with it off the link failed the same way before this change --
and by assembling bios/arch/m68k/intmask.S for both 68000 and ColdFire, and
the modified gemdosif.S, with m68k-linux-gnu-as.
